prints first 100 fibonacci

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -7,14 +7,14 @@
  */
 int main(void)
 {
-	long long int n1 = 1, n2 = 2;
-	long long int i, n3;
+	long int n1 = 1, n2 = 2;
+	long int i, n3;
 
-	printf("%lld, %lld, ", n1, n2);
+	printf("%ld, %ld, ", n1, n2);
 	for (i = 3; i < 101; i++)
 	{
 		n3 = n1 + n2;
-		printf("%lld", n3);
+		printf("%ld", n3);
 		if (i < 100)
 		{
 			printf(", ");
